Moved test mocks into own subfolder
