+++ __tests__/UnistToProseMirrorConverter.test.ts
@@ -3,7 +3,7 @@ import { Schema } from "prosemirror-model";
 
 import { ExtensionManager } from "../src/ExtensionManager";
 import { UnistToProseMirrorConverter } from "../src/UnistToProseMirrorConverter";
-import { MockSyntaxExtension } from "./MockSyntaxExtension";
+import { MockSyntaxExtension } from "./mocks/MockSyntaxExtension";
 
 jest.mock("../src/ExtensionManager");
 jest.mock("../src/SyntaxExtension");
